Jinsu (#32)
* UI 추가 및 수정, 에코 매장 좋아요 기능 추가

* feat: 내정보 챌린지 부분 오류 수정, 저장한 에코 매장 더보기 버튼 수정

+++ app/src/main/java/com/example/ecomate/ui/myprofile/MyProfileFragment.kt
@@ -95,7 +95,7 @@ class MyProfileFragment : Fragment() {
                             .into(challengeImage)
                     }
                     challengeName.text = it[0].challengeTitle
-                    challengeProgressBar.progress = (it[0].doneCnt / it[0].goalCnt) * 100
+                    challengeProgressBar.progress = ((it[0].doneCnt / it[0].goalCnt.toFloat()) * 100).toInt()
                     challengeProgressCount.text =
                         "${((it[0].doneCnt / it[0].goalCnt.toFloat()) * 100).toInt()}% 달성 (${it[0].doneCnt}회/${it[0].goalCnt}회)"
                 } else {
